+++ src/main/scala/co/topl/brambl/cli/views/WalletModelDisplayOps.scala
@@ -1,6 +1,12 @@
 package co.topl.brambl.cli.views
 
-import co.topl.brambl.dataApi.{WalletFellowship, WalletTemplate}
+import cats.Id
+import co.topl.brambl.builders.locks.LockTemplate
+import co.topl.brambl.builders.locks.PropositionTemplate
+import co.topl.brambl.dataApi.WalletFellowship
+import co.topl.brambl.dataApi.WalletTemplate
+import co.topl.brambl.utils.Encoding
+import io.circe.parser.parse
 
 object WalletModelDisplayOps {
 
@@ -13,7 +19,73 @@ object WalletModelDisplayOps {
   def display(walletEntity: WalletFellowship): String =
     s"""${walletEntity.xIdx}\t${walletEntity.name}"""
 
-  def display(walletTemplate: WalletTemplate): String =
-    s"""${walletTemplate.yIdx}\t${walletTemplate.name}\t${walletTemplate.lockTemplate}"""
+  def serialize[F[_]](lockTemplate: LockTemplate[F]): String = {
+    lockTemplate match {
+      case LockTemplate.PredicateTemplate(innerTemplates, threshold) =>
+        s"threshold($threshold, ${innerTemplates.map(serialize[F]).mkString(", ")})"
+    }
+  }
+  private def serialize[F[_]](lockTemplate: PropositionTemplate[F]): String = {
+    lockTemplate match {
+      case PropositionTemplate.OrTemplate(leftTemplate, rightTemplate) =>
+        val left = leftTemplate match {
+          case PropositionTemplate.AndTemplate(_, _) =>
+            s"(${serialize[F](leftTemplate)})"
+          case PropositionTemplate.OrTemplate(_, _) =>
+            s"(${serialize[F](leftTemplate)})"
+          case _ =>
+            serialize[F](leftTemplate)
+        }
+        val right = rightTemplate match {
+          case PropositionTemplate.AndTemplate(_, _) =>
+            s"(${serialize[F](rightTemplate)})"
+          case PropositionTemplate.OrTemplate(_, _) =>
+            s"(${serialize[F](rightTemplate)})"
+          case _ =>
+            serialize[F](rightTemplate)
+        }
+        s"$left or $right"
+      case PropositionTemplate.AndTemplate(leftTemplate, rightTemplate) =>
+        val left = leftTemplate match {
+          case PropositionTemplate.AndTemplate(_, _) =>
+            s"(${serialize[F](leftTemplate)})"
+          case PropositionTemplate.OrTemplate(_, _) =>
+            s"(${serialize[F](leftTemplate)})"
+          case _ =>
+            serialize[F](leftTemplate)
+        }
+        val right = rightTemplate match {
+          case PropositionTemplate.AndTemplate(_, _) =>
+            s"(${serialize[F](rightTemplate)})"
+          case PropositionTemplate.OrTemplate(_, _) =>
+            s"(${serialize[F](rightTemplate)})"
+          case _ =>
+            serialize[F](rightTemplate)
+        }
+        s"$left and $right"
+      case PropositionTemplate.ThresholdTemplate(innerTemplates, threshold) =>
+        s"threshold($threshold, ${innerTemplates.map(serialize[F]).mkString(", ")})"
+      case PropositionTemplate.SignatureTemplate(_, signIdx) =>
+        s"sign($signIdx)"
+      case PropositionTemplate.LockedTemplate(data) =>
+        s"locked(${data.map(x => Encoding.encodeToBase58(x.value.toByteArray())).getOrElse("")})"
+      case PropositionTemplate.HeightTemplate(_, min, max) =>
+        s"height($min, $max)"
+      case PropositionTemplate.TickTemplate(min, max) =>
+        s"tick($min, $max)"
+      case PropositionTemplate.DigestTemplate(_, digest) =>
+        s"digest(${Encoding.encodeToBase58(digest.value.toByteArray())})"
+    }
+  }
+
+  def display(walletTemplate: WalletTemplate): String = {
+    import co.topl.brambl.codecs.LockTemplateCodecs.decodeLockTemplate
+    (for {
+      json <- Id(parse(walletTemplate.lockTemplate))
+      decoded <- decodeLockTemplate[Id](json)
+    } yield s"""${walletTemplate.yIdx}\t\t${walletTemplate.name}\t\t${serialize(
+        decoded
+      )}""").toOption.get
+  }
 
 }